Create README.md
Bitcoins minining provides you with a trade to earn enough currency which will serve you for a while.
Furthermore let continue to trade in this appropriate mining trading which bring joy and happiness on our faces. And I will be happy to be ready to trade with any partner who want to trade to.